Added I2C support with configuration
Can configure the device to use a stop bit or not by sending 0x00 for no bit or 0x01 to a stop bit to the configuration characteristic.
Also moved the uart.h and uart.c files to a CustomSDK folder instead of Nordic SDK.  This is where the i2c.h and i2c.c files are as well.